hotfix for the bug causing an instance to be deleted wether the dialog result was confirmed or cancle

+++ client/src/app/dashboard/table-all/table-all.component.ts
@@ -77,13 +77,15 @@ export class TableAllComponent implements OnInit {
                 console.log('data', this.dataSource.data);
 
             } else {
-                this.apiService.deleteInstance(id).subscribe((deleteResult: HttpEvent<number>) => {
-                    console.log('result', deleteResult);
-                }, err => {
-                    console.log('error delete Instance');
-                });
-                this.removeAt(i);
-            }
+                if (result !== 'Cancle') {
+                    this.apiService.deleteInstance(id).subscribe((deleteResult: HttpEvent<number>) => {
+                        console.log('result', deleteResult);
+                    }, err => {
+                        console.log('error delete Instance');
+                    });
+                    this.removeAt(i);
+                }
+             }
             this.dialogResult = result;
         });
     }
